refactor(whatsrust): extract callback registrations
